Separated Point from all objects and removed BigDecimal from everywhere

+++ src/main/kotlin/com/euvsvirus/euvsvirus/workout/api/CreateWorkoutRequest.kt
@@ -1,11 +1,10 @@
 package com.euvsvirus.euvsvirus.workout.api
 
-import java.math.BigDecimal
 import java.time.ZonedDateTime
 
 data class CreateWorkoutRequest(
         val datetimeStart: ZonedDateTime,
         val datetimeEnd: ZonedDateTime,
         val sport: String,
-        val points: List<List<BigDecimal>>
+        val points: List<List<Double>>
 )
